RavenDB-17760 : address PR comments - mark TableSchema as partial class and move schema index definitions to a separate partial TableSchema - rename TableIndexType enum and its values
